refactor(ui): post-review cleanup — dead CSS, shared list row, token/a11y fixes
Acting on a multi-agent review of the redesign:
- Delete ~115 lines of orphaned CSS (.nd-action, .nd-prop-*, .nd-select,
  .rec-bar + keyframes) left behind when the redesign replaced them.
- Extract src/lib/noteList.ts + src/components/NoteListRow.tsx; AllNotes
  and Folder now share them instead of byte-identical copies.
- Unify the "selected" treatment: AllNotes filter chips + Settings
  theme/palette toggles use the soft-accent state (matching panel tabs);
  filled gold is reserved for primary actions.
- Add --color-warning-text so the admin role pill is legible in light
  mode (was ~2.4:1); point roleColorVar admin at it.
- Add aria-label to the overlay <select> pickers (were unlabeled).
- Sidebar collapse tooltip: drop the unwired "⌘\" claim.
- Add .nd-btn-icon-sm; route the inline icon buttons (panel close,
  regenerate) through .nd-btn-icon.
- Drop leftover var(--font-mono) (aliased to Hanken anyway).

Front-end only — no IPC/schema/behaviour changes. tsc clean; verified
light + dark in the app.
